src/requirements.txt
asgiref==3.2.7
boto3==1.14.11
botocore==1.17.11
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-appconf==1.0.4
django-cleanup==5.0.0
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-messages==0.6.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.10.0
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
